+++ my-new-app/src/app/app.component.html
@@ -3,4 +3,4 @@
 <!-- * * * * * * * * * * is only a placeholder * * * * * * * * * * -->
 <app-header></app-header>
 <!-- <app-main class="main"></app-main> -->
-<app-search-results [class]="'search-results'"></app-search-results>
+<router-outlet></router-outlet>
