+++ my-new-app/src/app/core/components/header/sorting/sorting.component.html
@@ -1,4 +1,4 @@
-<div class="sorting">
+<div [class]="'sorting'">
   <p>Sorting by:</p>
   <button appDateFilter>date</button>
   <button appViewFilter>count of views</button>
