+++ my-new-app/src/app/core/components/header/sorting/sorting.component.scss
@@ -1,10 +1,11 @@
-@import '../../../variablesSCSS.scss';
+@import '/src/app/variablesSCSS.scss';
 
 .sorting {
   display: flex;
   align-items: center;
   justify-content: center;
   font-size: small;
+  padding-top: 1vw;
   button {
     padding-left: .5vw;
     border-radius: .5vw;
@@ -14,6 +15,7 @@
     margin-left: .5vw;
   }
   p {
+    margin: 0;
     color: $sortingColor;
   }
 }
