+++ my-new-app/src/app/shared/components/search/search-item/search-item.component.html
@@ -1,13 +1,13 @@
 <div class="wrapper-card">
   <img src="{{ item.snippet.thumbnails.medium.url }}" alt="caption" />
   <div class="likes">
-    <img src="../../assets/svg/viewed.svg" alt="viewed" />
+    <img src="/src/assets/svg/viewed.svg" alt="viewed" />
     <p>{{ item.statistics.viewCount }}</p>
-    <img src="../../assets/svg/liked.svg" alt="liked" />
+    <img src="/src/assets/svg/liked.svg" alt="liked" />
     <p>{{ item.statistics.likeCount }}</p>
-    <img src="../../assets/svg/dislike.svg" alt="dislike" />
+    <img src="/src/assets/svg/dislike.svg" alt="dislike" />
     <p>{{ item.statistics.dislikeCount }}</p>
-    <img src="../../assets/svg/content-copy.svg" alt="coment" />
+    <img src="/src/assets/svg/content-copy.svg" alt="content" />
     <p>{{ item.statistics.commentCount }}</p>
   </div>
   <h3>Kuji Podcast #20</h3>
