+++ my-new-app/src/app/shared/components/search/search-item/search-item.component.scss
@@ -1,4 +1,4 @@
-@import '../../../variablesSCSS.scss';
+@import 'src/app/variablesSCSS.scss';
 
 .wrapper-card {
   display: flex;
