+++ my-new-app/src/app/shared/components/search/search.component.html
@@ -1,2 +1,2 @@
-<input class="search-input" type="text" placeholder="Search" />
-<button class="search-button" (click)="unloading()">search</button>
+<input [class]="'search-input'" type="text" placeholder="Search" />
+<button [class]="'search-button'" (click)="unloading()">search</button>
