+++ my-new-app/src/app/shared/components/search/search.component.scss
@@ -1,4 +1,4 @@
-@import '../../variablesSCSS.scss';
+@import 'src/app/variablesSCSS.scss';
 
 .search-input{
   width: 25vw;
