+++ my-new-app/src/app/shared/directives/date-filter.directive.ts
@@ -1,6 +1,6 @@
 import { Directive, ElementRef, HostListener, Renderer2 } from '@angular/core';
 import Data from 'src/Youtube-response/youtube-response';
-import { Item } from '../auth/models/search-item.model';
+import { Item } from '../../auth/models/search-item.model';
 import { FilterDateService } from '../services/filter-date.service';
 
 @Directive({
