+++ my-new-app/src/app/shared/directives/view-filter.directive.spec.ts
