+++ my-new-app/src/app/shared/services/filter-text.service.spec.ts
