+++ my-new-app/src/app/shared/services/filter-text.service.ts
@@ -4,12 +4,10 @@ import { Injectable, EventEmitter } from '@angular/core';
   providedIn: 'root'
 })
 export class FilterTextService {
-  // constructor() { }
+ 
   textService = '';
   onText:EventEmitter<string> = new EventEmitter();
    changeText(text: string){
     this.onText.emit(text);
-    // this.textService = text;
-    // console.log(this.textService)
   }
 }
