+++ my-new-app/src/app/shared/services/share-service.service.ts
