+++ my-new-app/src/app/youtube/pages/search-results/search-results.component.scss
