+++ my-new-app/src/app/youtube/youtube.module.ts
@@ -1,14 +1,18 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-
 import { YoutubeRoutingModule } from './youtube-routing.module';
+import { SearchResultsComponent } from './pages/search-results/search-results.component';
+import { SharedModule } from '../shared/shared.module';
 
 
 @NgModule({
-  declarations: [],
+  declarations: [
+    SearchResultsComponent,
+  ],
   imports: [
     CommonModule,
-    YoutubeRoutingModule
+    YoutubeRoutingModule,
+    SharedModule
   ]
 })
 export class YoutubeModule { }
